requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pillow
requests
python-dotenv
pygame
fastapi
uvicorn